requirements.txt
Django~=3.1.6

requests~=2.25.1
bs4~=0.0.1
beautifulsoup4~=4.9.3
djangorestframework~=3.12.2
python-firebase~=1.2